adding in test-cat, cleaning up changelog, fixes #106
