+++ aws/rs_aws_elb/aws_elb_plugin.rb
@@ -2,8 +2,8 @@ name 'aws_elb_plugin'
 type 'plugin'
 rs_ca_ver 20161221
 short_description "Amazon Web Services - Elastic Load Balancer"
-long_description "Version: 1.0"
-package "plugin/rs_aws_elb"
+long_description "Version: 1.1"
+package "plugins/rs_aws_elb"
 import "sys_log"
 
 plugin "rs_aws_elb" do
@@ -201,37 +201,6 @@ resource_pool "elb_pool" do
   end
 end
 
-parameter "lb_name" do
-  label "ELB Name"
-  description "ELB Name"
-  default "myelb-1"
-  type "string"
-end
-
-output "list_elb" do
-  label 'list action'
-end
-
-resource "my_elb", type: "rs_aws_elb.elb" do
-  name $lb_name
-  list_lbport "80"
-  list_instport "80"
-  list_proto "http"
-  list_instproto "http"
-  #subnet1 "subnet-7c295240"
-  #security_group1 "sg-a9b9e8d6"
-  az1 "us-east-1a"
-  az2 "us-east-1d"
-  #description "a simple elb"
-end
-
-operation 'list_elb' do
-  definition 'list_elbs'
-  output_mappings do{
-    $list_elb => $object
-  } end
-end
-
 define provision_elb(@declaration) return @elb do
   sub on_error: stop_debugging() do
     $object = to_object(@declaration)
@@ -252,17 +221,6 @@ define provision_elb(@declaration) return @elb do
     @elb = @elb.get()
     call stop_debugging()
   end
-end
-
-define list_elbs() return $object do
-  @elbs = rs_aws_elb.elb.empty()
-  sub on_error: stop_debugging() do
-    call start_debugging()
-    @elbs = rs_aws_elb.elb.list()
-    call stop_debugging()
-  end
-  $object = to_object(first(@elbs))
-  $object = to_s($object)
 end
 
 define delete_elb(@elb) do
