+++ aws/rs_aws_elb/changelog.md
@@ -1,5 +1,11 @@
 ELB Plugin changelog
 
+v1.1
+-----
+- adding Content-Type header and setting to application/xml
+- removing description from elb resource
+- creating test cat
+
 v1.0
 -----
 - initial release
